Displays segments correctly

+++ libstreetmap/src/ezgl/global.cpp
@@ -16,7 +16,7 @@ double avg_lat = 0;
 double avg_lon = 0;
 std::string userInput = {};
 std::vector<std::vector<LatLon>> points_on_segments;
-std::vector<std::vector<std::pair<float, float>>> xy_points_segments;
+std::vector<std::vector<std::pair<double, double>>> xy_points_segments;
 
 
 double x_from_lon(double lon) {
